enterprise-ai-assistant/requirements.txt
groq>=0.11.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
PyMuPDF>=1.24.0
pandas>=2.0.0
python-docx>=1.1.0
streamlit>=1.38.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
